// server/src/types.ts
export interface CourseInput {
    code: string;
    title: string;
    departmentId: number;
    description: string;
}